refactor: move adding dataset run item to ee folder
